Therapist Dashboard

+++ composeApp/src/commonMain/kotlin/domain/appointments/AppointmentRepository.kt
@@ -10,7 +10,7 @@ import domain.Models.TherapistAvailabilityResponse
 interface AppointmentRepository {
     suspend fun getAppointments(userId: Long, nextPage: Int = 1): Single<AppointmentListDataResponse>
     suspend fun postponeAppointment(appointment: Appointment, appointmentTime: Int,  day: Int, month: Int, year: Int): Single<ServerResponse>
-    suspend fun deleteAppointment(appointmentId: Int): Single<ServerResponse>
+    suspend fun deleteAppointment(appointmentId: Long): Single<ServerResponse>
     suspend fun joinMeeting(customParticipantId: String, presetName: String, meetingId: String): Single<JoinMeetingResponse>
     suspend fun getTherapistAvailability(therapistId: Int,vendorId: Long,day: Int, month: Int, year: Int): Single<TherapistAvailabilityResponse>
 }
